Introduces GitHub Actions workflow for Apache Cloudberry (incubating) build debian package
In commit cd0bfbc1ef10becce772022e40fec0196f28d974 was added docker build and test containers for ubuntu 22.04.

Here has been adding CI workflow to check if we still could create and install debian package for ubuntu 22.04.

To do this:

- use devops package for build cloudberry binary inside docker build container
- create debian package for builded binary
- upload artifacts to github actions
- download artifacts to ubuntu test container
- install downloaded debian packages
- check packages for integrity
- perform some installcheck tests (regress, pax, contrib, gpcontrib)

+++ devops/build/automation/cloudberry/scripts/configure-cloudberry.sh
@@ -104,6 +104,12 @@ CONFIGURE_LOG="${LOG_DIR}/configure.log"
 # Initialize environment
 init_environment "Cloudberry Configure Script" "${CONFIGURE_LOG}"
 
+# Check if BUILD_DESTINATION is set
+if [ -z "${BUILD_DESTINATION}" ]; then
+  log_completion "BUILD_DESTINATION is empty - error with exit"
+  exit 1
+fi
+
 # Initial setup
 log_section "Initial Setup"
 execute_cmd sudo rm -rf ${BUILD_DESTINATION} || exit 2
